requirements.txt
requests>=2.28.0
pydantic>=1.10.0
google-genai>=0.1.0 
sentence-transformers>=3.0.0
spacy>=3.7.0
holidays>=0.56
rapidfuzz>=3.9.0